[UI] Fix build break from constants consolidation
FILTER and PATTERN were dropped from utils/Enum.ts on the assumption
they had no importers, but utils/fileDownloader.ts (and its test)
imported them from there, breaking the Turbopack production build.
Point those imports at constants/navigator.ts, the constant's actual
canonical source post-consolidation, and restore navigator.ts's
PATTERN export, which was also mistakenly removed as unused.
